refactor(web): unify property value scope-priority resolution
The card, the detail-modal property list and PropertyContainer each
reimplemented "apply a per-resource preference over the global priority,
then walk the chain for the first non-empty scope" — four copies of two
algorithm halves, kept in sync by hand.

Extract buildEffectiveScopePriority / selectScopedValue / resolveScopedValue
into core/models/Resource.ts and route all three sites through them. The
fallbackOnEmpty chain truncation and empty-scope skip are unchanged; no
behavior change.

Closes #1149
